Downcasting of the offset for TextTrack removed (#398)
Summary: Downcasting of the offset for TextTrack removed
Type: Fix
Test Plan: UT/ CT, Fullstack
Jira: VPLAY-10419

+++ media/server/gstplayer/source/GstTextTrackSink.cpp
@@ -188,10 +188,10 @@ static GstFlowReturn gst_rialto_text_track_sink_render(GstBaseSink *sink, GstBuf
     if (gst_buffer_map(buffer, &info, GST_MAP_READ))
     {
         std::string data(reinterpret_cast<char *>(info.data), info.size);
-        int32_t displayOffset{0};
+        int64_t displayOffset{0};
         if (GST_BUFFER_OFFSET_NONE != GST_BUFFER_OFFSET(buffer))
         {
-            displayOffset = static_cast<int32_t>(GST_BUFFER_OFFSET(buffer));
+            displayOffset = static_cast<int64_t>(GST_BUFFER_OFFSET(buffer));
         }
         textTrackSink->priv->m_textTrackSession->sendData(data, 0 - displayOffset);
